backend/pyproject.toml
[tool.poetry]
name = "Bookscanner Backend"
version = "0.1.0"
description = "Backend for Bookscanner Application"
authors = ["reinhud"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
firebase-admin = "^6.5.0"
pillow = "^10.2.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
loguru = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-pyproject = "^1.2.3"
black = "^23.11.0"
pytest = "^7.4.3"
isort = "^5.13.2"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110

[tool.isort]
profile = "black"
